notifiy before delete resolved

+++ src/Pages/MyToysPage/MyToysCollection/MyToysCollection.jsx
@@ -8,30 +8,30 @@ const MyToysCollection = ({ toy, setReload, reload }) => {
 
     const handleDeleteToy = (id) => {
 
-        fetch(`https://car-world-server-henna.vercel.app/deleteToy/${id}`, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                Swal.fire({
-                    title: 'Are you sure?',
-                    text: "You won't be able to revert this!",
-                    icon: 'warning',
-                    showCancelButton: true,
-                    confirmButtonColor: '#3D4451',
-                    cancelButtonColor: '#d33',
-                    confirmButtonText: 'Yes, delete it!'
-                }).then((result) => {
-                    if (result.isConfirmed) {
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3D4451',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            fetch(`https://car-world-server-henna.vercel.app/deleteToy/${id}`, {
+                method: 'DELETE',
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
                         Swal.fire(
                             'Deleted!',
                             'Your file has been deleted.',
                             'success'
                         )
                     }
+                    setReload(!reload)
                 })
-                setReload(!reload)
-            })
+        })
     }
 
     return (
